Final code of the spike. We can filter on token with this code.
We can also control the use of a token or basic based on gemfire
properties provided on startup of the locator.

Further work is needed to ensure the filter chain will let traffic pass
when accessing endpoints that don't need secure access.

The solution assumes that provision of tokens happens outside the
responsibility of the Geode system such as when we use the CF CLI in
combination with UAA.

Co-authored-by: Joris Melchior <joris.melchior@gmail.com>
Co-authored-by: Darrel Schneider <dschneider@pivotal.io>